modify sim-trhepd-rheed; internal classes to separate files
